+++ src/environments/environment.ts
@@ -130,7 +130,7 @@ export class Environment {
   }
 
   public get futsHeaders(): HttpHeaders {
-    return this.headers.set(this.config?.featureFlags?.FEATURE_FLAG_NEW_API_ENDPOINTS ? 'x-fhir-profile' : 'fhirProfile', 'fhir-profile-snapshots');
+    return this.headers.set('x-fhir-profile', 'fhir-profile-snapshots');
   }
 
   public get featureFlags(): any {